travel_planner/requirements.txt
google-generativeai>=0.8.0,<1.0.0
langchain-google-genai>=2.0.0,<3.0.0
langchain-core>=0.3.0,<0.4.0
python-dotenv>=1.0.0,<2.0.0
tavily-python>=0.5.0,<1.0.0
streamlit>=1.40.0,<2.0.0
